Move buffer.cpp/h from base to Common
