+++ Common/Buffer.h
