+++ ext/native/net/http_client.cpp
@@ -17,16 +17,16 @@
 #endif
 
 #include <cmath>
-#include <stdio.h>
-#include <stdlib.h>
+#include <cstdio>
+#include <cstdlib>
 
-#include "base/buffer.h"
 #include "data/compression.h"
 #include "file/fd_util.h"
 #include "net/resolve.h"
 #include "net/url.h"
 #include "thread/threadutil.h"
 
+#include "Common/Buffer.h"
 #include "Common/Log.h"
 
 namespace net {
