+++ ext/native/net/http_client.h
@@ -5,7 +5,6 @@
 #include <thread>
 #include <cstdint>
 
-#include "base/buffer.h"
 #include "net/resolve.h"
 
 #ifdef _WIN32
@@ -20,6 +19,8 @@
 #include <sys/socket.h>
 #include <netdb.h>
 #endif
+
+#include "Common/Buffer.h"
 
 namespace net {
