+++ ext/native/net/http_server.cpp
@@ -34,12 +34,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "base/buffer.h"
 #include "file/fd_util.h"
 #include "net/http_server.h"
 #include "net/sinks.h"
 #include "thread/executor.h"
 
+#include "Common/Buffer.h"
 #include "Common/Log.h"
 
 namespace http {
